+++ packages/react-i18n/src/index.ts
@@ -1,13 +1,23 @@
-export {I18nManager, ExtractedTranslations} from './manager';
+export {I18nManager, ExtractedTranslations, RegisterOptions} from './manager';
 export {I18nContext} from './context';
-export {I18n} from './i18n';
+export {
+  CurrencyFormatOptions,
+  I18n,
+  NumberFormatOptions,
+  TranslateOptions,
+} from './i18n';
 export {useI18n} from './hooks';
 export {withI18n, WithI18nProps} from './decorator';
-export {memoizedNumberFormatter, translate} from './utilities';
+export {
+  memoizedNumberFormatter,
+  translate,
+  TranslateOptions as RootTranslateOptions,
+} from './utilities';
 export {
   I18nDetails,
   LanguageDirection,
   CurrencyCode,
+  Replacements,
   TranslationDictionary,
 } from './types';
 export {